+++ src/waveList.js
@@ -1,7 +1,8 @@
 import WaveSurfer from 'wavesurfer/wavesurfer.js';
 import TimelinePlugin from 'wavesurfer/plugin/timeline.js';
 import RegionPlugin from 'wavesurfer/plugin/regions.js';
-import fileDownloader from 'fileDownloader.js'
+import FileDownloader from 'fileDownloader.js'
+import FilterPlugin from 'filter.js'
 
 export default class WaveList {
     waveformId = 0;
@@ -139,7 +140,7 @@ export default class WaveList {
 
     bindLocalButtons(waveformNum, wsInstance) {
         $("#download" + waveformNum).click(function() {
-            fileDownloader.saveToWav(wsInstance.backend.buffer);
+            FileDownloader.saveToWav(wsInstance.backend.buffer);
         });
         $("#upload" + waveformNum).change(function() {
             wsInstance.loadBlob(this.files[0]);
@@ -196,164 +197,22 @@ export default class WaveList {
 
         $("#fade_in").unbind("click");
         $("#fade_in").click(function() {
-            if (this.currentRegionInfo != null) {
-                console.log(this.wavesurfers[this.currentRegionInfo.id].backend.buffer);
-                console.log(this.currentRegionInfo.region.start);
-                console.log(this.currentRegionInfo.region.end);
-
-                let selectedRegion = this.currentRegionInfo.region;
-                let selectedTrackBuffer = this.wavesurfers[this.currentRegionInfo.id].backend.buffer;
-
-                let audioLengthInSec = selectedTrackBuffer.duration;
-                let startPositionInSec = selectedRegion.start;
-                let endPositionInSec = Math.min(selectedRegion.end, audioLengthInSec);
-
-                if (startPositionInSec >= audioLengthInSec){
-                    window.alert("WARNING : Region is placed on outside of audio. [FADE-IN]");
-                    // Warning
-                }
-
-                else{
-                    let audioLengthInBuffer = selectedTrackBuffer.length;
-                    let startPositionInBuffer = parseInt(startPositionInSec / audioLengthInSec * audioLengthInBuffer);
-                    let endPositionInBuffer = parseInt(endPositionInSec / audioLengthInSec * audioLengthInBuffer);
-                    let regionLengthInBuffer = endPositionInBuffer - startPositionInBuffer;
-
-                    for (var channelNumber = 0; channelNumber < selectedTrackBuffer.numberOfChannels; channelNumber++){
-                        var channelData = selectedTrackBuffer.getChannelData(channelNumber);
-                        for (var cursor = startPositionInBuffer; cursor < endPositionInBuffer; cursor++){
-                            channelData[cursor] *= (cursor - startPositionInBuffer) / regionLengthInBuffer;
-                        }
-                    }
-                    this.wavesurfers[this.currentRegionInfo.id].drawer.fireEvent("redraw");
-                }
-            }
-            else {
-                window.alert("ERROR : Region not selected for operation [FADE-IN]");
-                // ERROR: User must specify fade-in region.
-                // Show error message to user.
-            }
+            FilterPlugin.fadeIn(this.currentRegionInfo, this.wavesurfers);
         }.bind(this));
 
         $("#fade_out").unbind("click");
         $("#fade_out").click(function() {
-            if (this.currentRegionInfo != null) {
-                console.log(this.wavesurfers[this.currentRegionInfo.id].backend.buffer);
-                console.log(this.currentRegionInfo.region.start);
-                console.log(this.currentRegionInfo.region.end);
-
-                let selectedRegion = this.currentRegionInfo.region;
-                let selectedTrackBuffer = this.wavesurfers[this.currentRegionInfo.id].backend.buffer;
-
-                let audioLengthInSec = selectedTrackBuffer.duration;
-                let startPositionInSec = selectedRegion.start;
-                let endPositionInSec = Math.min(selectedRegion.end, audioLengthInSec);
-
-                if (startPositionInSec >= audioLengthInSec){
-                    window.alert("WARNING : Region is placed on outside of audio. [FADE-OUT]");
-                    // Warning Concept
-                }
-
-                else{
-                    let audioLengthInBuffer = selectedTrackBuffer.length;
-                    let startPositionInBuffer = parseInt(startPositionInSec / audioLengthInSec * audioLengthInBuffer);
-                    let endPositionInBuffer = parseInt(endPositionInSec / audioLengthInSec * audioLengthInBuffer);
-                    let regionLengthInBuffer = endPositionInBuffer - startPositionInBuffer;
-
-                    for (var channelNumber = 0; channelNumber < selectedTrackBuffer.numberOfChannels; channelNumber++){
-                        var channelData = selectedTrackBuffer.getChannelData(channelNumber);
-                        for (var cursor = startPositionInBuffer; cursor < endPositionInBuffer; cursor++){
-                            channelData[cursor] *= (endPositionInBuffer - cursor) / regionLengthInBuffer;
-                        }
-                    }
-                    this.wavesurfers[this.currentRegionInfo.id].drawer.fireEvent("redraw");
-                }
-            }
-            else {
-                window.alert("ERROR : Region not selected for operation [FADE-OUT]");
-                // ERROR: User must specify trim region.
-                // Show error message to user.
-            }
+            FilterPlugin.fadeOut(this.currentRegionInfo, this.wavesurfers);
         }.bind(this));
 
         $("#trim").unbind("click");
         $("#trim").click(function() {
-            if (this.currentRegionInfo != null) {
-                console.log(this.wavesurfers[this.currentRegionInfo.id].backend.buffer);
-                console.log(this.currentRegionInfo.region.start);
-                console.log(this.currentRegionInfo.region.end);
-
-                let selectedRegion = this.currentRegionInfo.region;
-                let selectedTrackBuffer = this.wavesurfers[this.currentRegionInfo.id].backend.buffer;
-
-                let audioLengthInSec = selectedTrackBuffer.duration;
-                let startPositionInSec = selectedRegion.start;
-                let endPositionInSec = selectedRegion.end;
-
-                if (startPositionInSec >= audioLengthInSec){
-                    window.alert("ERROR : Region is placed on outside of audio. [TRIM]");
-                    // ERROR: Without this management,
-                    // Uncaught (in promise) DOMException: Unable to decode audio data
-                    // UNLESS : Try to make trim in void area.
-                }
-
-                else {
-                    let audioLengthInBuffer = selectedTrackBuffer.length;
-                    let startPositionInBuffer = startPositionInSec / audioLengthInSec * audioLengthInBuffer;
-                    let endPositionInBuffer = endPositionInSec / audioLengthInSec * audioLengthInBuffer;
-                    
-                    let blob = fileDownloader.saveToWav(selectedTrackBuffer, 
-                        startPositionInBuffer, endPositionInBuffer, true, this.wavesurfers[this.currentRegionInfo.id]);
-                    this.currentRegionInfo.region.remove();
-                }
-            }
-            else {
-                window.alert("ERROR : Region not selected for operation. [TRIM]");
-                // ERROR: User must specify trim region.
-                // Show error message to user.
-            }
+            FilterPlugin.trim(this.currentRegionInfo, this.wavesurfers);
         }.bind(this));
 
         $("#reverse").unbind("click");
         $("#reverse").click(function() {
-            if (this.currentRegionInfo != null) {
-                console.log(this.wavesurfers[this.currentRegionInfo.id].backend.buffer);
-                console.log(this.currentRegionInfo.region.start);
-                console.log(this.currentRegionInfo.region.end);
-
-                let selectedRegion = this.currentRegionInfo.region;
-                let selectedTrackBuffer = this.wavesurfers[this.currentRegionInfo.id].backend.buffer;
-
-                let audioLengthInSec = selectedTrackBuffer.duration;
-                let startPositionInSec = selectedRegion.start;
-                let endPositionInSec = Math.min(selectedRegion.end, audioLengthInSec);
-
-                if (startPositionInSec >= audioLengthInSec){
-                    window.alert("WARNING : Region is placed on outside of audio. [REVERSE]");
-                    // Warning Concept
-                }
-
-                else{
-                    let audioLengthInBuffer = selectedTrackBuffer.length;
-                    let startPositionInBuffer = parseInt(startPositionInSec / audioLengthInSec * audioLengthInBuffer);
-                    let endPositionInBuffer = parseInt(endPositionInSec / audioLengthInSec * audioLengthInBuffer);
-                    let regionLengthInBuffer = endPositionInBuffer - startPositionInBuffer;
-
-                    for (var channelNumber = 0; channelNumber < selectedTrackBuffer.numberOfChannels; channelNumber++){
-                        var channelData = selectedTrackBuffer.getChannelData(channelNumber);
-                        var cloneChannel = channelData.slice();
-                        for (var cursor = 0; cursor < regionLengthInBuffer; cursor++){
-                            channelData[startPositionInBuffer + cursor] = cloneChannel[endPositionInBuffer - 1 - cursor];
-                        }
-                    }
-                    this.wavesurfers[this.currentRegionInfo.id].drawer.fireEvent("redraw");
-                }
-            }
-            else {
-                window.alert("ERROR : Region not selected for operation [REVERSE]");
-                // ERROR: User must specify trim region.
-                // Show error message to user.
-            }
+            FilterPlugin.reverse(this.currentRegionInfo, this.wavesurfers);
         }.bind(this));
     }
